requirements.txt
python-dotenv
openai
python-docx